requirements.txt
Flask~=2.0.2
urllib3~=1.26.7
requests
flask_sqlalchemy
flask_login
flask-wtf
psycopg2
flask-migrate
flask_marshmallow
marshmallow-sqlalchemy
werkzeug
wtforms
sqlalchemy_utils
flask_classful
python-dotenv
flask-bootstrap
email_validator